1. Make the sample app dependent of the static library instead of the ENSDK.framework bundle for debugging and developing purposes 2. Remove the ENSDKUniversal target. We now use the build_framework.sh script inside 'scripts' folder to create the universal library Usage: cd scripts; ./build_framework.sh the script will generate a ENSDK.framework and copy the resource bundle we're using into the Products folder. 3rd party developers can grab the two bundles, add to their projects and they should be all set. 3. Now the private headers are put inside the framework under Headers/AdvancedHeaders. Developers can include the header like this: #include <ENSDK/AdvancedHeaders/EDAM.h> 4. Move all the assets in the Images.xcassets into a bundle, so we can access them in the client application. That's how Skitch did their integration.

+++ EvernoteSDKSample/ViewController.m
@@ -46,12 +46,12 @@
 
 #pragma mark - UITableView
 
-- (int)numberOfSectionsInTableView:(UITableView *)tableView
+- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView
 {
     return 1;
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     if ([[ENSession sharedSession] isAuthenticated]) {
         return 4; // Unauthenticate, user info, try activity, photo note
